requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
faiss-cpu==1.8.0
openai>=1.12.0
python-dotenv==1.0.0
aiofiles==23.2.0
numpy>=1.21.0,<2.0.0
jinja2>=3.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0